First drop...
This is a copy of the Arbotix/Trossen Robotics Bioloid classes, that I
did some slight extensions to, plus I modified the code to use standard
Arduino Serial(x) objects instead of custom code for the Atmega644p.
Also I was working with a few people up on Arduino to make some speed
enhancements to the the hardware serial class, to make this work well on
the 1.5x branch.  Not sure yet if the changes have made it into daily
builds and when scheduled for release.

In addition to this, I have made changes to the serial code here to make
it work with a Teensy 3.1.  With the Teensy I update the serial port to
be in half duplex mode and when needed changes the TX line from input to
output...

I have included a test program that I use for testing out stuff on my
Trossen Robotics PhantomX Hexapod and Quad, that has been modified to
use this library.